Explicitly stream a wide character formatted as a number.
In C++17, the underlying std::ostream can format only standard-width chars as characters, formatting wider characters of type char16_t or char32_t as numbers. In C++20, the overloads accepting these wide characters are explicitly deleted to avoid using such formatting unintentionally. Thus, this code no longer compiles.

Casting to integer types produces the same behavior in both versions as was produced in C++17 without the cast.

This change is part of a large-scale effort to migrate google3's default C++ version from 17 to 20.

PiperOrigin-RevId: 543893467

+++ src/config/character_form_manager.cc
@@ -414,7 +414,8 @@ void CharacterFormManagerImpl::SaveCharacterFormToStorage(
       storage_->Insert(group_key, reinterpret_cast<const char *>(&iform));
     }
   }
-  VLOG(2) << ucs2 << " is stored to " << kFileName << " as " << form;
+  VLOG(2) << static_cast<uint16_t>(ucs2) << " is stored to " << kFileName
+          << " as " << form;
 }
 
 void CharacterFormManagerImpl::ConvertString(const absl::string_view str,
